pkg/auth: move jwt to auth to be shared by syz-hub clients
