+++ pkg/auth/jwt.go
@@ -1,22 +1,26 @@
 // Copyright 2021 syzkaller project authors. All rights reserved.
 // Use of this source code is governed by Apache 2 LICENSE that can be found in the LICENSE file.
 
-package dashapi
+package auth
 
 import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"net/http"
 	"strings"
-	"sync"
 	"time"
 )
 
-type expiringToken struct {
-	token      string
-	expiration time.Time
+const (
+	DashboardAudience = "https://syzkaller.appspot.com/api"
+)
+
+type ExpiringToken struct {
+	Token      string
+	Expiration time.Time
 }
 
 // Returns the unverified expiration value from the given JWT token.
@@ -39,9 +43,12 @@ func extractJwtExpiration(token string) (time.Time, error) {
 	return time.Unix(claims.Expiration, 0), nil
 }
 
-// Queries the metadata server and returns the bearer token of the service account.
-// The token is scoped for the official dashboard.
-func retrieveJwtToken(ctor RequestCtor, doer RequestDoer) (*expiringToken, error) {
+// Queries the metadata server and returns the bearer token of the
+// service account. The token is scoped for the official dashboard.
+// The types of ctor and doer are the same as in http.NewRequest and
+// http.DefaultClient.Do.
+func RetrieveJwtToken(ctor func(method, url string, body io.Reader) (*http.Request, error),
+	doer func(req *http.Request) (*http.Response, error)) (*ExpiringToken, error) {
 	const v1meta = "http://metadata.google.internal/computeMetadata/v1"
 	req, err := ctor("GET", v1meta+"/instance/service-accounts/default/identity?audience="+DashboardAudience, nil)
 	if err != nil {
@@ -65,28 +72,5 @@ func retrieveJwtToken(ctor RequestCtor, doer RequestDoer) (*expiringToken, error
 	if err != nil {
 		return nil, err
 	}
-	return &expiringToken{token, expiration}, nil
-}
-
-// Augments the given doer with an authorization header carrying the
-// given token. The token gets refreshed when it becomes stale.
-func atachJwtToken(ctor RequestCtor, doer RequestDoer, token *expiringToken) RequestDoer {
-	lock := sync.Mutex{}
-	return func(req *http.Request) (*http.Response, error) {
-		lock.Lock()
-		if token.expiration.Before(time.Now()) {
-			// Keeping the lock while making http request is dubious, but
-			// making multiple concurrent requests is not any better.
-			t, err := retrieveJwtToken(ctor, doer)
-			if err != nil {
-				// Can't get a new token, so returning the error preemptively.
-				lock.Unlock()
-				return nil, err
-			}
-			*token = *t
-		}
-		req.Header.Add("Authorization", "Bearer "+token.token)
-		lock.Unlock()
-		return doer(req)
-	}
+	return &ExpiringToken{token, expiration}, nil
 }
